pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "linguaclaw"
version = "0.1.0"
description = "A Claude-Code-style agent harness implemented mostly in natural language"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["agent harness", "agents", "llm", "nlah", "linguaclaw", "runtime"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "litellm>=1.79.0",
  "python-dotenv>=1.0.1",
  "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
  "build>=1.2.2",
  "twine>=6.1.0",
]

[project.urls]
Homepage = "https://github.com/curated-skills/NLAH"
Repository = "https://github.com/curated-skills/NLAH"
Documentation = "https://github.com/curated-skills/NLAH#readme"
Issues = "https://github.com/curated-skills/NLAH/issues"
Paper = "https://arxiv.org/abs/2603.25723"

[project.scripts]
linguaclaw = "linguaclaw:main"

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["linguaclaw"]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["appserver", "appserver.*", "rich_cli", "rich_cli.*"]
